requirements.txt
chardet==5.2.0
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
fonttools==4.55.2
fpdf2==2.8.2
grapheme==0.6.0
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.0.0
PyMuPDF==1.25.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.16.0
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
reportlab==4.3.1
scipy==1.14.1
six==1.17.0
tesseract==0.1.3
tzdata==2024.2
